LayoutTests/imported/w3c/web-platform-tests/css/css-sizing/contain-intrinsic-size/contain-intrinsic-size-logical-003-expected.txt


PASS .test 1
PASS .test 2
PASS .test 3
PASS .test 4
PASS .test 5
PASS .test 6
PASS .test 7
PASS .test 8
PASS .test 9
PASS .test 10
PASS .test 11
PASS .test 12
PASS .test 13
PASS .test 14
PASS .test 15
PASS .test 16
PASS .test 17
PASS .test 18
PASS .test 19
PASS .test 20
PASS .test 21
PASS .test 22
PASS .test 23
PASS .test 24
FAIL .test 25 assert_equals:
<canvas class="test cis-none" width="40" height="20" data-expected-client-width="0" data-expected-client-height="0"></canvas>
clientWidth expected 0 but got 769
FAIL .test 26 assert_equals:
<canvas class="test cis-block" width="40" height="20" data-expected-client-width="0" data-expected-client-height="50"></canvas>
clientWidth expected 0 but got 100
FAIL .test 27 assert_equals:
<canvas class="test cis-inline" width="40" height="20" data-expected-client-width="100" data-expected-client-height="0"></canvas>
clientHeight expected 0 but got 50
PASS .test 28
FAIL .test 29 assert_equals:
<canvas class="test cis-none vertical" width="40" height="20" data-expected-client-height="0" data-expected-client-width="0"></canvas>
clientWidth expected 0 but got 1538
FAIL .test 30 assert_equals:
<canvas class="test cis-block vertical" width="40" height="20" data-expected-client-height="0" data-expected-client-width="50"></canvas>
clientHeight expected 0 but got 25
FAIL .test 31 assert_equals:
<canvas class="test cis-inline vertical" width="40" height="20" data-expected-client-height="100" data-expected-client-width="0"></canvas>
clientWidth expected 0 but got 200
PASS .test 32
PASS .test 33
PASS .test 34
PASS .test 35
PASS .test 36
PASS .test 37
PASS .test 38
PASS .test 39
PASS .test 40
PASS .test 41
PASS .test 42
PASS .test 43
PASS .test 44
PASS .test 45
PASS .test 46
PASS .test 47
PASS .test 48
